# 8341127: Extra call to MethodHandle::asType from memory segment var handles fails to inline