The fix for JDK-8331865 introduced an accidental performance regression.
The main issue is that now *all* memory segment var handles go through some round of adaptation.
Adapting a var handle results in a so called *indirect* var handle.
When an indirect var handle is called through a *var handle guard*, an extra `MethodHandle::asType` call is triggered.
In some cases, if `asType` has already been compiled into a big method, it cannot be inlined into the caller, which then results in a failure to inline through the var handle call, resulting in a big performance regression.

The solution is to make sure that the compiled size of `MethodHandle::asType` stays small: this is done by making sure that the slowpath (the one which populates the cache used by `asType`) is not inlined by the JVM. This is done by consolidating the slow path into a separate method, which is annotated with the internal `@DontInline` annotation.

This problem was originally reported here:
https://mail.openjdk.org/pipermail/panama-dev/2024-September/020643.html

While we did not test this fix directly, we have made sure that running the problematic benchmark with the flags:

```
-XX:CompileCommand=dontinline,java/lang/invoke/MethodHandle.setAsTypeCache
-XX:CompileCommand=dontinline,java/lang/invoke/MethodHandle.asTypeUncached
```

Solves the performance regression. The fix in this PR is just a programmatic way to achieve the same results w/o the need of command line flags.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8341127](https://bugs.openjdk.org/browse/JDK-8341127): Extra call to MethodHandle::asType from memory segment var handles fails to inline (**Enhancement** - P2)


### Reviewers
 * [Claes Redestad](https://openjdk.org/census#redestad) (@cl4es - **Reviewer**) Review applies to [f644d52d](https://git.openjdk.org/jdk/pull/21283/files/f644d52defecf82f9405046f981aa69a2176a0a3)
 * [Jorn Vernee](https://openjdk.org/census#jvernee) (@JornVernee - **Reviewer**) Review applies to [964a273f](https://git.openjdk.org/jdk/pull/21283/files/964a273f73636b7a45417d30f2756403b0c21323)
 * [Paul Sandoz](https://openjdk.org/census#psandoz) (@PaulSandoz - **Reviewer**)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/21283/head:pull/21283` \
`$ git checkout pull/21283`

Update a local copy of the PR: \
`$ git checkout pull/21283` \
`$ git pull https://git.openjdk.org/jdk.git pull/21283/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 21283`

View PR using the GUI difftool: \
`$ git pr show -t 21283`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/21283.diff">https://git.openjdk.org/jdk/pull/21283.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/21283#issuecomment-2385386109)