@mcimadamore 2024-10-01
Benchmark results

Before:

```
Benchmark                               (asTypeCompiled)  Mode  Cnt  Score   Error  Units
LoopOverNonConstantAsType.segment_loop             false  avgt   30  0.232 ± 0.003  ms/op
LoopOverNonConstantAsType.segment_loop              true  avgt   30  6.943 ± 0.100  ms/op
LoopOverNonConstantAsType.unsafe_loop              false  avgt   30  0.223 ± 0.002  ms/op
LoopOverNonConstantAsType.unsafe_loop               true  avgt   30  0.227 ± 0.002  ms/op
```

After:

```
Benchmark                               (asTypeCompiled)  Mode  Cnt  Score   Error  Units
LoopOverNonConstantAsType.segment_loop             false  avgt   30  0.237 ± 0.002  ms/op
LoopOverNonConstantAsType.segment_loop              true  avgt   30  0.233 ± 0.003  ms/op
LoopOverNonConstantAsType.unsafe_loop              false  avgt   30  0.233 ± 0.002  ms/op
LoopOverNonConstantAsType.unsafe_loop               true  avgt   30  0.234 ± 0.002  ms/op
```