@mcimadamore 2024-10-01
Alternatives considered:
1. don't inline slow path (implemented in this PR)
2. add `@ForceInline` to `MethodHandle::asType`
3. add extra check on var handle guards to skip redundant `asType` calls

Of these, (1) was preferred as it fixes the underlying issue (and not just for memory segment var handles), while at the same time avoiding unwanted side-effects (e.g. adding `@ForceInline` can often result in too much code getting compiled, which can then create issues with other callee-side thresholds).