@mcimadamore 2024-10-01
> LGTM
> 
> In the micro perhaps the `unsafe_loop` variant is superfluous for this special case test? At least it's distinct from the test in `LoopOverNonConstant` by summing longs rather than ints.

Yeah - that is a bit redundant - however could be useful to have some sort of baseline, so I left it in there.