@iwanowww 2024-10-01
I don't have a clear understanding how `VarHandleGuards` is used, but I assume that `MethodHandle::asType` calls there are no-ops in scenarios you care about.

It does make sense to annotate`setAsTypeCache` with `@DontInline`, but I'd also add `@ForceInline` on `asType`.