@mcimadamore 2024-10-04
/backport :jdk23